Move manager related tests to proper module
